+++ subsys/bluetooth/audio/CMakeLists.txt
@@ -4,7 +4,7 @@ add_subdirectory_ifdef(CONFIG_BT_SHELL shell)
 
 zephyr_library()
 
-zephyr_library_sources_ifdef(CONFIG_BT_CONN audio.c)
+zephyr_library_sources(audio.c)
 
 if (CONFIG_BT_VOCS OR CONFIG_BT_VOCS_CLIENT)
 	zephyr_library_sources(vocs.c)
